+++ django_comments_threaded/admin.py
@@ -5,7 +5,7 @@ from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
 from django.contrib import admin
-from django_comments_threaded.models import Comment, LastReaded
+from django_comments_threaded.models import Comment, LastRead
 from mptt.admin import MPTTModelAdmin
 
 
@@ -20,4 +20,4 @@ class LastReadAdmin(admin.ModelAdmin):
 
 
 admin.site.register(Comment, CommentAdmin)
-admin.site.register(LastReaded, LastReadedAdmin)
+admin.site.register(LastRead, LastReadAdmin)