+++ django_comments_threaded/plugins.py
@@ -4,7 +4,7 @@ from __future__ import unicode_literals
 from __future__ import print_function
 from __future__ import absolute_import
 from __future__ import division
-from django.conf.urls import patterns, url
+from django.conf.urls import url
 from django_comments.plugins import BasePlugin
 from django_comments_threaded.models import Comment
 from django_comments_threaded.forms import CommentForm, CommentReplyForm
@@ -26,7 +26,11 @@ class ThreadedCommentPlugin(BasePlugin):
 
     def get_urlpatterns(self):
         return [
-            url('^create/$',            CreateView.as_view(plugin=self), name='threaded_comments_create'), 
-            url('^reply/(?P<pk>\d+)/$', CreateView.as_view(plugin=self), name='threaded_comments_reply'),
-            url('^update/(?P<content_type>\d+)/(?P<object_pk>\d+)/$', UpdateCommentsView.as_view(plugin=self), name='threaded_comments_update'),
+            url('^create/$', CreateView.as_view(plugin=self),
+                name='threaded_comments_create'),
+            url('^reply/(?P<pk>\d+)/$', CreateView.as_view(plugin=self),
+                name='threaded_comments_reply'),
+            url('^update/(?P<content_type>\d+)/(?P<object_pk>\d+)/$',
+                UpdateCommentsView.as_view(plugin=self),
+                name='threaded_comments_update'),
         ]