+++ tests.py
@@ -4,7 +4,6 @@
 from __future__ import unicode_literals
 from __future__ import absolute_import
 from django.conf import settings
-from django import get_version
 import sys
 import os